Revert back to previous state

+++ commands/clear.js
@@ -4,13 +4,10 @@ module.exports = {
     args: false,
     extra: false,
     execute(message, args) {
-        if (!message.channel.parent && !message.channel.nsfw) return;
-        if ((message.channel.parent.name !== bots) || (!message.channel.nsfw)) return;
-
         message.channel.fetchMessages({ limit: 100 })
             .then(messages => {
                 amount = `${messages.size}`;
-                if (amount >= 2 && amount <= 100) {
+                if (amount >= 2 && amount <= 100 && (message.channel.parent.name === "bots")) {
                   message.channel
                     .bulkDelete(amount, true)
                     .catch(err => {
